% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_model.R
\name{block_model}
\alias{block_model}
\title{Crear un modelo de bloques}
\usage{
block_model(drillhole, coordinates, dim, df)
}
\arguments{
\item{drillhole}{corresponde a la base de datos con las coordenadas y la variable a graficar}

\item{coordinates}{corresponde a los nombres de las coordenadas}

\item{dim}{es la dimension de cada bloque}

\item{pot}{es el porcentaje de la distancia maxima a la cual se incluiran los bloques cercanos  a los sondajes}
}
\description{
Esta funcion permite crear un modelo de bloques en base a sondajes 2D o 3D. El resultado
corresponde a los centroides de los bloques que se encuentran a una cierta distancia de los sondajes
}
\details{
Nada
}
