% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_drillhole.R
\name{composite_drillhole}
\alias{composite_drillhole}
\title{Compositar sondajes}
\usage{
composite_drillhole(drillhole, composite)
}
\arguments{
\item{drillhole}{corresponde a una salida de la funcion import_drillhole}

\item{composite}{corresponde al tamano del composito}
}
\description{
Esta funcion permite compositar sondajes
}
\details{
Esta funcion sirve para compositar sondajes
}
