% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expvar_dir.R
\name{expvar_dir}
\alias{expvar_dir}
\title{Calcular el variograma experimental direccional}
\usage{
expvar_dir(
  drillhole,
  coordinates,
  variable,
  azimuth,
  dip,
  lags,
  maxdist,
  tol.lag,
  tol.dir
)
}
\arguments{
\item{drillhole}{corresponde a la base de datos con las coordenadas y la variable a utilizar}

\item{coordinates}{corresponde a los nombres de las coordenadas}

\item{variable}{corresponde al nombre de la variable a utilizar}

\item{azimuth}{corresponde a los azimuth de las direcciones a utilizar}

\item{dip}{corresponde a los dip de las direcciones a utilizar}

\item{lags}{es el numero de lags del variograma experimental}

\item{maxdist}{es la maxima distancia a la cual se calcula el variograma}

\item{tol.lag}{corresponde a la tolerancia porcentual en el lag de los variogramas}

\item{tol.dir}{corresponde a la tolerancia angular en la direccion de los variogramas}
}
\description{
Esta funcion permite calcular el variograma experimental direccional
}
\details{
Esta funcion sirve para sondajes 2D o 3D
}
