% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expvar_omni.R
\name{expvar_omni}
\alias{expvar_omni}
\title{Calcular el variograma experimental omnidireccional}
\usage{
expvar_omni(drillhole, coordinates, variable, lags, maxdist, tol.lag)
}
\arguments{
\item{drillhole}{corresponde a la base de datos con las coordenadas y la variable a utilizar}

\item{coordinates}{corresponde a los nombres de las coordenadas}

\item{variable}{corresponde al nombre de la variable a utilizar}

\item{lags}{es el numero de lags del variograma experimental}

\item{maxdist}{es la maxima distancia a la cual se calcula el variograma}

\item{tol.lag}{corresponde a la tolerancia porcentual en el lag de los variogramas}
}
\description{
Esta funcion permite calcular el variograma experimental omnidireccional
}
\details{
Esta funcion sirve para sondajes 2D o 3D
}
