% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitvar_dir.R
\name{fitvar_dir}
\alias{fitvar_dir}
\title{Ajustar el variograma experimental direccional}
\usage{
fitvar_dir(variograma, lags, pares, modelos, nug)
}
\arguments{
\item{variograma}{corresponde a los variogramas experimentales direccional}

\item{lags}{corresponde a los lags de los variogramas entregados}

\item{pares}{corresponde al numero de pares de cada lag de los variogramas direccional}

\item{modelos}{corresponde a los modelos teoricos a ajustar a los variogramas experimentales.
Los modelos permitidos son "Pepa","Exponencial","Esferico" y "Gaussiano".}

\item{nug}{corresponde a un valor de verdadero o falso, indicando si se incluira una
pepa en los variogramas a ajustar o no.}
}
\description{
Esta funcion permite ajustar el variograma experimental direccional
}
\details{
Nada
}
