% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitvar_omni.R
\name{fitvar_omni}
\alias{fitvar_omni}
\title{Ajustar el variograma experimental omnidireccional}
\usage{
fitvar_omni(variograma, lags, pares, modelo, nug)
}
\arguments{
\item{variograma}{corresponde al variograma experimental omnidireccional}

\item{lags}{corresponde a los lags del variograma entregado}

\item{pares}{corresponde al numero de pares de cada lag del variograma omnidireccional}

\item{modelo}{corresponde al modelo teorico a ajustar al variograma experimental.
Los modelos permitidos son "Pepa","Exponencial","Esferico" y "Gaussiano".}

\item{nug}{corresponde a un valor de verdadero o falso, indicando si se incluira una
pepa en el variograma a ajustar o no.}
}
\description{
Esta funcion permite ajustar el variograma experimental omnidireccional
}
\details{
Nada
}
