% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging_dir.R
\name{kriging_dir}
\alias{kriging_dir}
\title{Kriging direccional 2D y 3D}
\usage{
kriging_dir(
  sondajes,
  coordenadas.sondajes,
  variable,
  localidades,
  coordenadas.localidades,
  kriging,
  azimuth,
  dip,
  variogramas,
  pepa,
  mesetas,
  alcances,
  media,
  max.num.pares
)
}
\arguments{
\item{sondajes}{corresponde a la base de datos con las coordenadas y la variable a utilizar}

\item{coordenadas.sondajes}{corresponde a los nombres de las coordenadas en los sondajes}

\item{variable}{corresponde al nombre de la variable a utilizar}

\item{localidades}{corresponde a los centroides del modelo de bloques a construir}

\item{coordenadas.localidades}{corresponde a los nombres de las coordenadas del modelo de bloques}

\item{kriging}{corresponde al tipo de Kriging a realizar, el que puede ser "Simple" o "Ordinario"}

\item{azimuth}{corresponde a los azimuth de los variogramas direccionales}

\item{dip}{corresponde a los dip de los variogramas direccionales, solo en el caso 3D}

\item{variogramas}{corresponde a los modelos de variogramas a utilizar, los que pueden ser "Exponencial", "Esferico" o "Gaussiano"}

\item{pepa}{corresponde a la pepa de los variogramas a utilizar para la prediccion}

\item{media}{corresponde a la media de la variable en caso de realizar Kriging Simple}

\item{max.num.pares}{corresponde al numero maximo de pares para realizar la prediccion}

\item{meseta}{corresponde a las mesetas de los variogramas a utilizar para la prediccion}

\item{alcance}{corresponde a los alcances de los variogramas a utilizar para la prediccion}
}
\description{
Esta funcion permite hacer prediccion por medio de kriging direccional, para muestras 2D y 3D
}
\details{
Esta funcion sirve para sondajes 2D o 3D
}
