% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_drillhole.R
\name{plot_drillhole}
\alias{plot_drillhole}
\title{Graficar sondajes}
\usage{
plot_drillhole(drillhole, coordinates, variable, pt.type, pt.size, pt.col)
}
\arguments{
\item{drillhole}{corresponde a la base de datos con las coordenadas y la variable a graficar}

\item{coordinates}{corresponde a los nombres de las coordenadas}

\item{variable}{corresponde al nombre de la variable a graficar}

\item{pt.type}{es la clase de punto en el grafico}

\item{pt.size}{es el tamano del punto en el grafico}

\item{pt.col}{es el color del punto en el grafico}
}
\description{
Esta funcion permite graficar sondajes 2D o 3D usando o no una
variable continua o categorica en base a una escala de colores predefinida
}
\details{
En el caso de usar una variable continua el grafico mostrara ademas
de los sondajes un histograma con la densidad de la variable. En el caso de una
variable categorica mostrara un grafico de torta con las clases y sus porcentajes.
En el caso de no definir una variable se graficaran solamente las posiciones de los sondajes.
}
