% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_section.R
\name{plot_section}
\alias{plot_section}
\title{Graficar secciones de sondajes}
\usage{
plot_section(
  drillhole,
  coordinates,
  variable,
  plane,
  level,
  tol,
  pt.type,
  pt.size,
  pt.col
)
}
\arguments{
\item{drillhole}{corresponde a la base de datos con las coordenadas y la variable a graficar}

\item{coordinates}{corresponde a los nombres de las coordenadas}

\item{variable}{corresponde al nombre de la variable a graficar}

\item{plane}{corresponde a la dimension del plano a graficar, XY, XZ o YZ}

\item{level}{corresponde a la ubicacion del plano}

\item{tol}{corresponde a un parametro que permite capturar mas datos dentro del plano}

\item{pt.type}{es la clase de punto en el grafico}

\item{pt.size}{es el tamano del punto en el grafico}

\item{pt.col}{es el color del punto en el grafico}
}
\description{
Esta funcion permite graficar secciones 2D de un conjunto de sondajes
usando o no una variable continua o categorica en base a una escala de colores predefinida
}
\details{
En el caso de no definir un level se graficara el punto medio de la dimension de interes
}
