% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotvar_dir.R
\name{plotvar_dir}
\alias{plotvar_dir}
\title{Graficar el variograma experimental direccional}
\usage{
plotvar_dir(expvar, pares, fitvar)
}
\arguments{
\item{expvar}{un objeto proveniente de la funcion "expvar_dir"}

\item{pares}{un valor TRUE o FALSE, indicando si se desea agregar el numero de pares de cada lag en el grafico}

\item{fitvar}{un objeto proveniente de la funcion "fitvar_dir". Se entrega en el
caso de querer agregar el variograma teorico al grafico}
}
\description{
Esta funcion permite graficar el variograma experimental direccional
y ademas incorporar a el un variograma teorico
}
\details{
Nada
}
