% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotvar_omni.R
\name{plotvar_omni}
\alias{plotvar_omni}
\title{Graficar el variograma experimental omnidireccional}
\usage{
plotvar_omni(expvar, pares, fitvar)
}
\arguments{
\item{expvar}{un objeto proveniente de la funcion "expvar_omni"}

\item{pares}{un valor TRUE o FALSE, indicando si se desea agregar el numero de pares de cada lag en el grafico}

\item{fitvar}{un objeto proveniente de la funcion "fitvar_omni". Se entrega en el
caso de querer agregar el variograma teorico al grafico}
}
\description{
Esta funcion permite graficar el variograma experimental omnidireccional
y ademas incorporar a el un variograma teorico
}
\details{
Nada
}
