% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_omni.R
\name{valid_omni}
\alias{valid_omni}
\title{Validacion cruzada omnidireccional 2D y 3D}
\usage{
valid_omni(
  sondajes,
  coordenadas.sondajes,
  variable,
  kriging,
  variograma,
  pepa,
  meseta,
  alcance,
  media,
  max.num.pares
)
}
\arguments{
\item{sondajes}{corresponde a la base de datos con las coordenadas y la variable a utilizar}

\item{coordenadas.sondajes}{corresponde a los nombres de las coordenadas en los sondajes}

\item{variable}{corresponde al nombre de la variable a utilizar}

\item{kriging}{corresponde al tipo de Kriging a realizar, el que puede ser "Simple" o "Ordinario"}

\item{variograma}{corresponde al modelo de variograma a utilizar, el que puede ser "Exponencial", "Esferico" o "Gaussiano"}

\item{pepa}{corresponde a la pepa del variograma a utilizar para la prediccion}

\item{meseta}{corresponde a la meseta del variograma a utilizar para la prediccion}

\item{alcance}{corresponde al alcance del variograma a utilizar para la prediccion}

\item{media}{corresponde a la media de la variable en caso de realizar Kriging Simple}

\item{max.num.pares}{corresponde al numero maximo de vecinos para realizar la prediccion}
}
\description{
Esta funcion permite realizar validacion cruzada por medio de kriging omnidireccional, para muestras 2D y 3D
}
\details{
Esta funcion sirve para sondajes 2D o 3D
}
