% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varmap.R
\name{varmap}
\alias{varmap}
\title{Calcular el mapa variografico}
\usage{
varmap(drillhole, coordinates, variable, lags, maxdist)
}
\arguments{
\item{drillhole}{corresponde a la base de datos con las coordenadas y la variable a utilizar}

\item{coordinates}{corresponde a los nombres de las coordenadas}

\item{variable}{corresponde al nombre de la variable a utilizar}

\item{lags}{es el numero de lags del mapa variografico}

\item{maxdist}{es la maxima distancia a la cual se calcula el mapa variografico}
}
\description{
Esta funcion permite calcular el mapa variografico experimental direccional
}
\details{
Esta funcion sirve para crear un mapa variografico 2D o 3D
}
